Add files via upload
On average Kenya tests about 5000 people from all the counties. The tested people all from ages 1 -80
yrs old.
At the end of the day the ministry of health gives a report on the following:
1. Total number of tests done
2. How many tested positive.
3. Distribution of positive cases according to counties
4. Distribution of positive cases according to age ( less than 5, 5 to 30, 30-40, 50-60, 60-80)
It is required a program that can simulate the testing of covid-19 in Kenya and afterwards give the
results as outlined above.